move MakeRandomString to strutil
