+++ oauth/oauth.go
@@ -24,7 +24,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/ubuntu-core/snappy/helpers"
+	"github.com/ubuntu-core/snappy/strutil"
 )
 
 // Token contains the sso token
@@ -72,7 +72,7 @@ func quote(s string) string {
 func MakePlaintextSignature(token *Token) string {
 	// hrm, rfc5849 says that nonce, timestamp are not used for PLAINTEXT
 	// but our sso server is unhappy without, so
-	nonce := helpers.MakeRandomString(60)
+	nonce := strutil.MakeRandomString(60)
 	timestamp := time.Now().Unix()
 
 	s := fmt.Sprintf(`OAuth oauth_nonce="%s", oauth_timestamp="%v", oauth_version="1.0", oauth_signature_method="PLAINTEXT", oauth_consumer_key="%s", oauth_token="%s", oauth_signature="%s%%26%s"`, nonce, timestamp, quote(token.ConsumerKey), quote(token.TokenKey), quote(token.ConsumerSecret), quote(token.TokenSecret))
